+++ src/main/scala/io/pg/webhook/webhook.scala
@@ -3,8 +3,7 @@ package io.pg.webhook
 import cats.data.NonEmptyList
 import sttp.tapir.server.ServerEndpoint
 import cats.implicits._
-import io.circe.Codec
-import io.pg.webhook.WebhookRouter.transport.WebhookEvent
+import io.pg.gitlab.transport.WebhookEvent
 
 object WebhookRouter {
 
@@ -17,45 +16,6 @@ object WebhookRouter {
     NonEmptyList.of(
       webhook.serverLogicRecoverErrors(b => ("Successful! Got this hot body: " + b).pure[F])
     )
-  }
-
-  import io.circe.generic.extras._
-  import io.circe.generic.extras.semiauto._
-
-  implicit val config: Configuration =
-    Configuration.default.withSnakeCaseMemberNames.withSnakeCaseConstructorNames
-
-  object transport {
-    final case class WebhookEvent(project: Project, objectKind: WebhookEvent.ObjectKind)
-
-    object WebhookEvent {
-
-      sealed trait ObjectKind extends Product with Serializable
-
-      object ObjectKind {
-        case object Push extends ObjectKind
-        case object MergeRequest extends ObjectKind
-
-        implicit val codec: Codec[ObjectKind] = deriveEnumerationCodec
-      }
-
-      implicit val codec: Codec[WebhookEvent] = deriveConfiguredCodec
-
-    }
-
-    final case class Project(
-      id: Int /* todo: apparently tapir has a conflict in schemas for Long */,
-      name: String,
-      pathWithNamespace: String,
-      defaultBranch: String,
-      url: String
-    )
-
-    object Project {
-
-      implicit val codec: Codec[Project] = deriveConfiguredCodec
-    }
-
   }
 
 }
